fix: typos and renamings
- fixed "PodcasstEpisodes" typo;
- renamed getEpisodeDuration to useEpisodeDuration;
