+++ src/app/(pages)/podcast-episodes/[slug]/page.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import React from "react";
 
 import { notFound } from "next/navigation";
